+++ backend/canisters/group/impl/src/model/mod.rs
@@ -1,3 +1,2 @@
 pub mod activity_notification_state;
-pub mod invited_users;
 pub mod new_joiner_rewards;