+++ backend/canisters/group/CHANGELOG.md
@@ -9,6 +9,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Fix bug where user could be not found for c2c calls ([#3712](https://github.com/open-ic/open-chat/pull/3712))
 
+### Changed
+
+- Move invited_users from Data to GroupChatCore ([#3714](https://github.com/open-ic/open-chat/pull/3714))
+
 ## [[2.0.705](https://github.com/open-ic/open-chat/releases/tag/v2.0.705-group)] - 2023-06-01
 
 ### Added